cheri: Convert sysent probe parameters to intptr_t
For pure-capability ABIs we need to pass capabilities not just
addresses.

Effort:		CHERI upstreaming
Sponsored by:	Innovate UK

+++ sys/tools/syscalls/scripts/systrace_args.lua
@@ -43,9 +43,9 @@ function systrace_args.generate(tbl, config, fh)
 
 	gen:write(string.format([[
 static void
-systrace_args(int sysnum, void *params, uint64_t *uarg, int *n_args)
+systrace_args(int sysnum, void *params, uintptr_t *uarg, int *n_args)
 {
-	int64_t *iarg = (int64_t *)uarg;
+	intptr_t *iarg = (intptr_t *)uarg;
 	int a = 0;
 	switch (sysnum) {
 ]]))
